python/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
mistralai
presidio-analyzer
presidio-anonymizer
spacy
python-dotenv
requests
pandas
openpyxl
python-docx
